requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
diskcache==5.6.3
filelock==3.18.0
fsspec==2025.7.0
huggingface-hub==0.34.4
idna==3.10
Jinja2==3.1.6
llama_cpp_python @ file:///C:/Users/Hyde/Documents/ComfyAI_latest/llama-cpp-python
MarkupSafe==3.0.2
numpy==2.3.2
packaging==25.0
PyYAML==6.0.2
requests==2.32.4
tqdm==4.67.1
typing_extensions==4.14.1
urllib3==2.5.0
